requirements.txt
Flask==2.1.2
Flask_RESTful==0.3.9
mock==4.0.3
pandas==1.4.2
Pillow==9.2.0
plantuml==0.3.0
psycopg2==2.9.3
requests==2.27.1
streamlit==1.8.1
